move robot dialog rendering to new module
